*** empty log message ***
git-svn-id: svn://svn.code.sf.net/p/iup/iup/trunk@41 2f08bc20-984a-4df8-b195-9b5d5d477537

+++ iup/src/win/iupwin_globalattrib.c
@@ -123,7 +123,7 @@ char *iupdrvGetGlobal(const char *name)
     EnumDisplayMonitors(NULL, NULL, winMonitorInfoEnum, (LPARAM)monitors_rect);
 
     for (i=0; i < monitors_count; i++)
-      pstr += sprintf(pstr, "%d %d %d %d\n", monitors_rect[i].left, monitors_rect[i].top, monitors_rect[i].right-monitors_rect[i].left, monitors_rect[i].bottom-monitors_rect[i].top);
+      pstr += sprintf(pstr, "%d %d %d %d\n", (int)monitors_rect[i].left, (int)monitors_rect[i].top, (int)(monitors_rect[i].right-monitors_rect[i].left), (int)(monitors_rect[i].bottom-monitors_rect[i].top));
 
     free(monitors_rect);
     return str;
